:sparkles: Api, type 정리 + 종료된 공모주 디자인 수정
